requirements.txt
Flask==2.2.5
tensorflow==2.13.0
numpy==1.24.3
pillow==10.0.0
